style: Move font-family outside of mako
