+++ components/script/dom/element.rs
@@ -587,9 +587,9 @@ impl LayoutElementHelpers for LayoutDom<Element> {
                 shared_lock,
                 PropertyDeclaration::FontFamily(
                     font_family::SpecifiedValue::Values(
-                        font_family::computed_value::FontFamilyList::new(vec![
-                            font_family::computed_value::FontFamily::from_atom(
-                                font_family)])))));
+                        computed::font::FontFamilyList::new(Box::new([
+                            computed::font::SingleFontFamily::from_atom(
+                                font_family)]))))));
         }
 
         let font_size = self.downcast::<HTMLFontElement>().and_then(|this| this.get_size());
