+++ components/style/properties/longhand/font.mako.rs
@@ -7,536 +7,12 @@
 
 <% data.new_style_struct("Font", inherited=True) %>
 
-<%def name="nongecko_unreachable()">
-    %if product == "gecko":
-        ${caller.body()}
-    %else:
-        unreachable!()
-    %endif
-</%def>
-
-<%helpers:longhand name="font-family" animation_value_type="discrete"
-                   flags="APPLIES_TO_FIRST_LETTER APPLIES_TO_FIRST_LINE APPLIES_TO_PLACEHOLDER"
-                   spec="https://drafts.csswg.org/css-fonts/#propdef-font-family">
-    #[cfg(feature = "gecko")] use gecko_bindings::bindings;
-    #[cfg(feature = "gecko")] use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
-    use properties::longhands::system_font::SystemFont;
-    use self::computed_value::{FontFamily, FontFamilyList, FamilyName};
-    use std::fmt;
-    use style_traits::ToCss;
-
-
-    pub mod computed_value {
-        use cssparser::{CssStringWriter, Parser, serialize_identifier};
-        #[cfg(feature = "gecko")] use gecko_bindings::{bindings, structs};
-        #[cfg(feature = "gecko")] use gecko_bindings::sugar::refptr::RefPtr;
-        #[cfg(feature = "gecko")] use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
-        use std::fmt::{self, Write};
-        #[cfg(feature = "gecko")] use std::hash::{Hash, Hasher};
-        #[cfg(feature = "servo")] use std::slice;
-        use style_traits::{ToCss, ParseError};
-        use Atom;
-        pub use self::FontFamily as SingleComputedValue;
-
-        #[derive(Clone, Debug, Eq, Hash, MallocSizeOf, PartialEq)]
-        #[cfg_attr(feature = "servo", derive(Deserialize, Serialize))]
-        pub enum FontFamily {
-            FamilyName(FamilyName),
-            Generic(Atom),
-        }
-
-        #[derive(Clone, Debug, Eq, Hash, MallocSizeOf, PartialEq)]
-        #[cfg_attr(feature = "servo", derive(Deserialize, Serialize))]
-        pub struct FamilyName {
-            pub name: Atom,
-            pub syntax: FamilyNameSyntax,
-        }
-
-        #[derive(Clone, Debug, Eq, Hash, MallocSizeOf, PartialEq)]
-        #[cfg_attr(feature = "servo", derive(Deserialize, Serialize))]
-        pub enum FamilyNameSyntax {
-            /// The family name was specified in a quoted form, e.g. "Font Name"
-            /// or 'Font Name'.
-            Quoted,
-
-            /// The family name was specified in an unquoted form as a sequence of
-            /// identifiers.  The `String` is the serialization of the sequence of
-            /// identifiers.
-            Identifiers(String),
-        }
-
-        impl FontFamily {
-            #[inline]
-            pub fn atom(&self) -> &Atom {
-                match *self {
-                    FontFamily::FamilyName(ref family_name) => &family_name.name,
-                    FontFamily::Generic(ref name) => name,
-                }
-            }
-
-            #[inline]
-            #[cfg(not(feature = "gecko"))] // Gecko can't borrow atoms as UTF-8.
-            pub fn name(&self) -> &str {
-                self.atom()
-            }
-
-            #[cfg(not(feature = "gecko"))] // Gecko can't borrow atoms as UTF-8.
-            pub fn from_atom(input: Atom) -> FontFamily {
-                match input {
-                    atom!("serif") |
-                    atom!("sans-serif") |
-                    atom!("cursive") |
-                    atom!("fantasy") |
-                    atom!("monospace") => {
-                        return FontFamily::Generic(input)
-                    }
-                    _ => {}
-                }
-                match_ignore_ascii_case! { &input,
-                    "serif" => return FontFamily::Generic(atom!("serif")),
-                    "sans-serif" => return FontFamily::Generic(atom!("sans-serif")),
-                    "cursive" => return FontFamily::Generic(atom!("cursive")),
-                    "fantasy" => return FontFamily::Generic(atom!("fantasy")),
-                    "monospace" => return FontFamily::Generic(atom!("monospace")),
-                    _ => {}
-                }
-
-                // We don't know if it's quoted or not. So we set it to
-                // quoted by default.
-                FontFamily::FamilyName(FamilyName {
-                    name: input,
-                    syntax: FamilyNameSyntax::Quoted,
-                })
-            }
-
-            /// Parse a font-family value
-            pub fn parse<'i, 't>(input: &mut Parser<'i, 't>) -> Result<Self, ParseError<'i>> {
-                if let Ok(value) = input.try(|i| i.expect_string_cloned()) {
-                    return Ok(FontFamily::FamilyName(FamilyName {
-                        name: Atom::from(&*value),
-                        syntax: FamilyNameSyntax::Quoted,
-                    }))
-                }
-                let first_ident = input.expect_ident()?.clone();
-
-                // FIXME(bholley): The fast thing to do here would be to look up the
-                // string (as lowercase) in the static atoms table. We don't have an
-                // API to do that yet though, so we do the simple thing for now.
-                let mut css_wide_keyword = false;
-                match_ignore_ascii_case! { &first_ident,
-                    "serif" => return Ok(FontFamily::Generic(atom!("serif"))),
-                    "sans-serif" => return Ok(FontFamily::Generic(atom!("sans-serif"))),
-                    "cursive" => return Ok(FontFamily::Generic(atom!("cursive"))),
-                    "fantasy" => return Ok(FontFamily::Generic(atom!("fantasy"))),
-                    "monospace" => return Ok(FontFamily::Generic(atom!("monospace"))),
-                    % if product == "gecko":
-                        "-moz-fixed" => return Ok(FontFamily::Generic(atom!("-moz-fixed"))),
-                    % endif
-
-                    // https://drafts.csswg.org/css-fonts/#propdef-font-family
-                    // "Font family names that happen to be the same as a keyword value
-                    //  (`inherit`, `serif`, `sans-serif`, `monospace`, `fantasy`, and `cursive`)
-                    //  must be quoted to prevent confusion with the keywords with the same names.
-                    //  The keywords ‘initial’ and ‘default’ are reserved for future use
-                    //  and must also be quoted when used as font names.
-                    //  UAs must not consider these keywords as matching the <family-name> type."
-                    "inherit" => css_wide_keyword = true,
-                    "initial" => css_wide_keyword = true,
-                    "unset" => css_wide_keyword = true,
-                    "default" => css_wide_keyword = true,
-                    _ => {}
-                }
-
-                let mut value = first_ident.as_ref().to_owned();
-                let mut serialization = String::new();
-                serialize_identifier(&first_ident, &mut serialization).unwrap();
-
-                // These keywords are not allowed by themselves.
-                // The only way this value can be valid with with another keyword.
-                if css_wide_keyword {
-                    let ident = input.expect_ident()?;
-                    value.push(' ');
-                    value.push_str(&ident);
-                    serialization.push(' ');
-                    serialize_identifier(&ident, &mut serialization).unwrap();
-                }
-                while let Ok(ident) = input.try(|i| i.expect_ident_cloned()) {
-                    value.push(' ');
-                    value.push_str(&ident);
-                    serialization.push(' ');
-                    serialize_identifier(&ident, &mut serialization).unwrap();
-                }
-                Ok(FontFamily::FamilyName(FamilyName {
-                    name: Atom::from(value),
-                    syntax: FamilyNameSyntax::Identifiers(serialization),
-                }))
-            }
-
-            #[cfg(feature = "gecko")]
-            /// Return the generic ID for a given generic font name
-            pub fn generic(name: &Atom) -> (structs::FontFamilyType, u8) {
-                use gecko_bindings::structs::FontFamilyType;
-                if *name == atom!("serif") {
-                    (FontFamilyType::eFamily_serif,
-                     structs::kGenericFont_serif)
-                } else if *name == atom!("sans-serif") {
-                    (FontFamilyType::eFamily_sans_serif,
-                     structs::kGenericFont_sans_serif)
-                } else if *name == atom!("cursive") {
-                    (FontFamilyType::eFamily_cursive,
-                     structs::kGenericFont_cursive)
-                } else if *name == atom!("fantasy") {
-                    (FontFamilyType::eFamily_fantasy,
-                     structs::kGenericFont_fantasy)
-                } else if *name == atom!("monospace") {
-                    (FontFamilyType::eFamily_monospace,
-                     structs::kGenericFont_monospace)
-                } else if *name == atom!("-moz-fixed") {
-                    (FontFamilyType::eFamily_moz_fixed,
-                     structs::kGenericFont_moz_fixed)
-                } else {
-                    panic!("Unknown generic {}", name);
-                }
-            }
-
-            #[cfg(feature = "gecko")]
-            fn from_font_family_name(family: &structs::FontFamilyName) -> FontFamily {
-                use gecko_bindings::structs::FontFamilyType;
-
-                match family.mType {
-                    FontFamilyType::eFamily_serif => FontFamily::Generic(atom!("serif")),
-                    FontFamilyType::eFamily_sans_serif => FontFamily::Generic(atom!("sans-serif")),
-                    FontFamilyType::eFamily_monospace => FontFamily::Generic(atom!("monospace")),
-                    FontFamilyType::eFamily_cursive => FontFamily::Generic(atom!("cursive")),
-                    FontFamilyType::eFamily_fantasy => FontFamily::Generic(atom!("fantasy")),
-                    FontFamilyType::eFamily_moz_fixed => FontFamily::Generic(Atom::from("-moz-fixed")),
-                    FontFamilyType::eFamily_named => {
-                        let name = Atom::from(&*family.mName);
-                        let mut serialization = String::new();
-                        serialize_identifier(&name.to_string(), &mut serialization).unwrap();
-                        FontFamily::FamilyName(FamilyName {
-                            name: name.clone(),
-                            syntax: FamilyNameSyntax::Identifiers(serialization),
-                        })
-                    },
-                    FontFamilyType::eFamily_named_quoted => FontFamily::FamilyName(FamilyName {
-                        name: (&*family.mName).into(),
-                        syntax: FamilyNameSyntax::Quoted,
-                    }),
-                    x => panic!("Found unexpected font FontFamilyType: {:?}", x),
-                }
-            }
-        }
-
-        impl ToCss for FamilyName {
-            fn to_css<W>(&self, dest: &mut W) -> fmt::Result where W: fmt::Write {
-                match self.syntax {
-                    FamilyNameSyntax::Quoted => {
-                        dest.write_char('"')?;
-                        write!(CssStringWriter::new(dest), "{}", self.name)?;
-                        dest.write_char('"')
-                    }
-                    FamilyNameSyntax::Identifiers(ref serialization) => {
-                        // Note that `serialization` is already escaped/
-                        // serialized appropriately.
-                        dest.write_str(&*serialization)
-                    }
-                }
-            }
-        }
-
-        impl ToCss for FontFamily {
-            fn to_css<W>(&self, dest: &mut W) -> fmt::Result where W: fmt::Write {
-                match *self {
-                    FontFamily::FamilyName(ref name) => name.to_css(dest),
-
-                    // All generic values accepted by the parser are known to not require escaping.
-                    FontFamily::Generic(ref name) => {
-                        % if product == "gecko":
-                            // We should treat -moz-fixed as monospace
-                            if name == &atom!("-moz-fixed") {
-                                return dest.write_str("monospace");
-                            }
-                        % endif
-
-                        write!(dest, "{}", name)
-                    },
-                }
-            }
-        }
-
-        impl ToCss for T {
-            fn to_css<W>(&self, dest: &mut W) -> fmt::Result where W: fmt::Write {
-                let mut iter = self.0.iter();
-                iter.next().unwrap().to_css(dest)?;
-                for family in iter {
-                    dest.write_str(", ")?;
-                    family.to_css(dest)?;
-                }
-                Ok(())
-            }
-        }
-
-        #[cfg(feature = "servo")]
-        #[derive(Clone, Debug, Eq, Hash, MallocSizeOf, PartialEq)]
-        pub struct FontFamilyList(Vec<FontFamily>);
-
-        #[cfg(feature = "gecko")]
-        #[derive(Clone, Debug)]
-        pub struct FontFamilyList(pub RefPtr<structs::SharedFontList>);
-
-        #[cfg(feature = "gecko")]
-        impl Hash for FontFamilyList {
-            fn hash<H>(&self, state: &mut H) where H: Hasher {
-                for name in self.0.mNames.iter() {
-                    name.mType.hash(state);
-                    name.mName.hash(state);
-                }
-            }
-        }
-
-        #[cfg(feature = "gecko")]
-        impl PartialEq for FontFamilyList {
-            fn eq(&self, other: &FontFamilyList) -> bool {
-                if self.0.mNames.len() != other.0.mNames.len() {
-                    return false;
-                }
-                for (a, b) in self.0.mNames.iter().zip(other.0.mNames.iter()) {
-                    if a.mType != b.mType || &*a.mName != &*b.mName {
-                        return false;
-                    }
-                }
-                true
-            }
-        }
-
-        #[cfg(feature = "gecko")]
-        impl Eq for FontFamilyList {}
-
-        impl FontFamilyList {
-            #[cfg(feature = "servo")]
-            pub fn new(families: Vec<FontFamily>) -> FontFamilyList {
-                FontFamilyList(families)
-            }
-
-            #[cfg(feature = "gecko")]
-            pub fn new(families: Vec<FontFamily>) -> FontFamilyList {
-                let fontlist;
-                let names;
-                unsafe {
-                    fontlist = bindings::Gecko_SharedFontList_Create();
-                    names = &mut (*fontlist).mNames;
-                    names.ensure_capacity(families.len());
-                };
-
-                for family in families {
-                    match family {
-                        FontFamily::FamilyName(ref f) => {
-                            let quoted = matches!(f.syntax, FamilyNameSyntax::Quoted);
-                            unsafe {
-                                bindings::Gecko_nsTArray_FontFamilyName_AppendNamed(
-                                    names,
-                                    f.name.as_ptr(),
-                                    quoted
-                                );
-                            }
-                        }
-                        FontFamily::Generic(ref name) => {
-                            let (family_type, _generic) = FontFamily::generic(name);
-                            unsafe {
-                                bindings::Gecko_nsTArray_FontFamilyName_AppendGeneric(
-                                    names,
-                                    family_type
-                                );
-                            }
-                        }
-                    }
-                }
-
-                FontFamilyList(unsafe { RefPtr::from_addrefed(fontlist) })
-            }
-
-            #[cfg(feature = "servo")]
-            pub fn iter(&self) -> slice::Iter<FontFamily> {
-                self.0.iter()
-            }
-
-            #[cfg(feature = "gecko")]
-            pub fn iter(&self) -> FontFamilyNameIter {
-                FontFamilyNameIter {
-                    names: &self.0.mNames,
-                    cur: 0,
-                }
-            }
-
-            #[cfg(feature = "gecko")]
-            /// Return the generic ID if it is a single generic font
-            pub fn single_generic(&self) -> Option<u8> {
-                let mut iter = self.iter();
-                if let Some(FontFamily::Generic(ref name)) = iter.next() {
-                    if iter.next().is_none() {
-                        return Some(FontFamily::generic(name).1);
-                    }
-                }
-                None
-            }
-        }
-
-        #[cfg(feature = "gecko")]
-        pub struct FontFamilyNameIter<'a> {
-            names: &'a structs::nsTArray<structs::FontFamilyName>,
-            cur: usize,
-        }
-
-        #[cfg(feature = "gecko")]
-        impl<'a> Iterator for FontFamilyNameIter<'a> {
-            type Item = FontFamily;
-
-            fn next(&mut self) -> Option<Self::Item> {
-                if self.cur < self.names.len() {
-                    let item = FontFamily::from_font_family_name(&self.names[self.cur]);
-                    self.cur += 1;
-                    Some(item)
-                } else {
-                    None
-                }
-            }
-        }
-
-        #[derive(Clone, Debug, Eq, Hash, PartialEq)]
-        #[cfg_attr(feature = "servo", derive(MallocSizeOf))]
-        pub struct T(pub FontFamilyList);
-
-        #[cfg(feature = "gecko")]
-        impl MallocSizeOf for T {
-            fn size_of(&self, _ops: &mut MallocSizeOfOps) -> usize {
-                // SharedFontList objects are generally shared from the pointer
-                // stored in the specified value. So only count this if the
-                // SharedFontList is unshared.
-                unsafe {
-                    bindings::Gecko_SharedFontList_SizeOfIncludingThisIfUnshared(
-                        (self.0).0.get()
-                    )
-                }
-            }
-        }
-    }
-
-    #[inline]
-    pub fn get_initial_value() -> computed_value::T {
-        computed_value::T(
-            FontFamilyList::new(vec![FontFamily::Generic(atom!("serif"))])
-        )
-    }
-
-    /// <family-name>#
-    /// <family-name> = <string> | [ <ident>+ ]
-    /// TODO: <generic-family>
-    pub fn parse<'i, 't>(_: &ParserContext, input: &mut Parser<'i, 't>)
-                         -> Result<SpecifiedValue, ParseError<'i>> {
-        SpecifiedValue::parse(input)
-    }
-
-    #[derive(Clone, Debug, Eq, Hash, PartialEq)]
-    pub enum SpecifiedValue {
-        Values(FontFamilyList),
-        System(SystemFont),
-    }
-
-    #[cfg(feature = "gecko")]
-    impl SpecifiedValue {
-        /// Return the generic ID if it is a single generic font
-        pub fn single_generic(&self) -> Option<u8> {
-            match *self {
-                SpecifiedValue::Values(ref values) => values.single_generic(),
-                _ => None,
-            }
-        }
-    }
-
-    impl ToComputedValue for SpecifiedValue {
-        type ComputedValue = computed_value::T;
-        fn to_computed_value(&self, _cx: &Context) -> Self::ComputedValue {
-            match *self {
-                SpecifiedValue::Values(ref v) => computed_value::T(v.clone()),
-                SpecifiedValue::System(_) => {
-                    <%self:nongecko_unreachable>
-                        _cx.cached_system_font.as_ref().unwrap().font_family.clone()
-                    </%self:nongecko_unreachable>
-                }
-            }
-        }
-        fn from_computed_value(other: &computed_value::T) -> Self {
-            SpecifiedValue::Values(other.0.clone())
-        }
-    }
-
-    #[cfg(feature = "gecko")]
-    impl MallocSizeOf for SpecifiedValue {
-        fn size_of(&self, _ops: &mut MallocSizeOfOps) -> usize {
-            match *self {
-                SpecifiedValue::Values(ref v) => {
-                    // Although a SharedFontList object is refcounted, we always
-                    // attribute its size to the specified value.
-                    unsafe {
-                        bindings::Gecko_SharedFontList_SizeOfIncludingThis(
-                            v.0.get()
-                        )
-                    }
-                }
-                SpecifiedValue::System(_) => 0,
-            }
-        }
-    }
-
-    impl SpecifiedValue {
-        pub fn system_font(f: SystemFont) -> Self {
-            SpecifiedValue::System(f)
-        }
-        pub fn get_system(&self) -> Option<SystemFont> {
-            if let SpecifiedValue::System(s) = *self {
-                Some(s)
-            } else {
-                None
-            }
-        }
-
-        pub fn parse<'i, 't>(input: &mut Parser<'i, 't>) -> Result<Self, ParseError<'i>> {
-            input.parse_comma_separated(|input| FontFamily::parse(input)).map(|v| {
-                SpecifiedValue::Values(FontFamilyList::new(v))
-            })
-        }
-    }
-
-    impl ToCss for SpecifiedValue {
-        fn to_css<W>(&self, dest: &mut W) -> fmt::Result where W: fmt::Write {
-            match *self {
-                SpecifiedValue::Values(ref v) => {
-                    let mut iter = v.iter();
-                    iter.next().unwrap().to_css(dest)?;
-                    for family in iter {
-                        dest.write_str(", ")?;
-                        family.to_css(dest)?;
-                    }
-                    Ok(())
-                }
-                SpecifiedValue::System(sys) => sys.to_css(dest),
-            }
-        }
-    }
-
-    /// `FamilyName::parse` is based on `FontFamily::parse` and not the other way around
-    /// because we want the former to exclude generic family keywords.
-    impl Parse for FamilyName {
-        fn parse<'i, 't>(_: &ParserContext, input: &mut Parser<'i, 't>) -> Result<Self, ParseError<'i>> {
-            match FontFamily::parse(input) {
-                Ok(FontFamily::FamilyName(name)) => Ok(name),
-                Ok(FontFamily::Generic(_)) => Err(input.new_custom_error(StyleParseErrorKind::UnspecifiedError)),
-                Err(e) => Err(e)
-            }
-        }
-    }
-</%helpers:longhand>
+${helpers.predefined_type("font-family",
+                          "FontFamily",
+                          initial_value="computed::FontFamily::serif()",
+                          animation_value_type="discrete",
+                          flags="APPLIES_TO_FIRST_LETTER APPLIES_TO_FIRST_LINE APPLIES_TO_PLACEHOLDER",
+                          spec="https://drafts.csswg.org/css-fonts/#propdef-font-family")}
 
 ${helpers.single_keyword_system("font-style",
                                 "normal italic oblique",
@@ -833,7 +309,7 @@ ${helpers.predefined_type("-x-text-zoom",
                 use gecko_bindings::bindings;
                 use gecko_bindings::structs::{LookAndFeel_FontID, nsFont};
                 use std::mem;
-                use values::computed::font::FontSize;
+                use values::computed::font::{FontSize, FontFamilyList};
 
                 let id = match *self {
                     % for font in system_fonts:
@@ -855,7 +331,7 @@ ${helpers.predefined_type("-x-text-zoom",
                 let weight = longhands::font_weight::computed_value::T::from_gecko_weight(system.weight);
                 let ret = ComputedSystemFont {
                     font_family: longhands::font_family::computed_value::T(
-                        longhands::font_family::computed_value::FontFamilyList(
+                        FontFamilyList(
                             unsafe { system.fontlist.mFontlist.mBasePtr.to_safe() }
                         )
                     ),
